semana_8/requirements.txt
flask
werkzeug
pyjwt
pytest
